+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-11 03:25:52
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 20d249c5152dd7c97dfd042e31e6e24e8916e4b5
+
+## [ForwardWidgets-2kuai] - 2025-08-11 03:25:52
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: da2058b94aaf57190641f472bf6ed67ade16e946
+
+## [ForwardWidgets-pack1r] - 2025-08-11 03:25:52
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 29ace0bb1772ccffd1792b0e96ccc168a25c67a2
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-10 03:27:08
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main